Made LARUL work with VxWorks/WPILib and Diab
Thus far, most of this code has not actually been compiled… TILL NOW. I
just got all of it to compile under Diab in a standard robot project,
and had to made so many changes, i don’t know if i can accurately tell
you what. Basically it comes down to three things: Naming conflicts
with WPILib (oops), VxWorks fuckery (literally the worst attempt at a
POSIX wrapper i’ve ever had the displeasure to work with) and the
leniency of GCC when not compiling to a strict spec. gross.